update for new output, which is sometimes truncated...
the "create raw attribute" code seems to create things like

raw.parent1.child = 0xxxx

instead of

raw.parent.child.1 = 0xxxxx

as a result, when it's encoded, the "raw" portion of the data
is truncated.  This can be investigated later, as it's not
a show stopper.

See also the wimax.txt tests, which have a few @todo comments
noticing this same issue